Created 2 numpy code files on random and one serial,
the random is the numpy code also vectorised this need more looking at, the serial has a slightly different version to the normal vectorised numpy code. Needs to be tested

I also added a base target, i think this needs to be added to testing as im doing it manually currently. The randomisation here is making it tricky